move request tests closer to component
